+++ src/utils/custom-types.ts
@@ -12,20 +12,3 @@ export type ErrorInputProps = {
   email: string;
   password: string;
 };
-
-export type ExerciseType = {
-  exerciseId: string;
-  _id: string;
-};
-
-export type WorkoutProps = {
-  _id: string;
-  name: string;
-  description: string;
-  duration: number;
-  difficulty: number;
-  exercises: ExerciseType[];
-  athleteIds: number[];
-  authorId: number;
-  __v: number;
-};
